Move Book & Genre classes to models.py
